backend-python/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
pillow
scikit-image
scipy
